chore: move bufferize to module level functions
